QCFold-main/QCFold-main/requirements.txt
torch>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
biopython>=1.81
pennylane>=0.35.0
pennylane-lightning>=0.35.0
pyyaml>=6.0
matplotlib>=3.7.0
seaborn>=0.12.0
pandas>=2.0.0
scikit-learn>=1.3.0
tqdm>=4.65.0
